requirements.txt
asgiref==3.7.2
blinker==1.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
command-not-found==0.3
crispy-bootstrap5==0.7
cryptography==41.0.4
dbus-python==1.2.18
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.2.6
django-allauth==0.57.0
django-auth-ldap==4.6.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-embed-video==1.4.9
django-filter==23.5
django-ldap-sync==0.5.0
django-multiselectfield==0.1.12
django-rest-knox==4.2.0
djangorestframework==3.14.0
filelock==3.6.0
gunicorn @ git+https://github.com/benoitc/gunicorn.git@792edf6d9aabcbfb84e76be1d722ac49c32dc027
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Markdown==3.5.1
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.2
Pillow==10.1.0
platformdirs==2.5.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyGObject==3.42.1
PyJWT==2.8.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-decouple==3.8
python-ldap==3.4.3
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==5.4.1
requests==2.31.0
requests-oauthlib==1.3.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
systemd-python==234
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.7
virtualenv==20.13.0+ds
wadllib==1.3.6
whitenoise==6.6.0
zipp==1.0.0
